refactor: single --fields choke point + peel content/publish studio routes
Keep applyFieldsFilter only in cli.ts so UNKNOWN_FIELD drives exit codes;
formatOutput formats already-filtered results. Extract /api/cmo/content/*
and /api/publish* handlers into server/routes/{content,publish}.ts and wire
lib/schemas.ts as SSOT for ACTIVITY/BRAND (+ content/publish) body schemas.
